Changed my OBJ loader implementation to use the tinyobjloader instead, which is way faster in Release mode and more reliable. Will probably hook up the materials as well as differentiating between objects inside a mesh. The next step will be integrating Dear ImGUI as the primary GUI system until mine is ready for use. This will be the foundation to building the next big project, the PrehistoricEditor

+++ PrehistoricEngine/src/engine/platform/opengl/rendering/shaders/environmentMap/GLPrefilterShader.cpp
@@ -24,6 +24,6 @@ namespace Prehistoric
 
 		SetUniformf("roughness", roughness);
 		SetUniformf("resolution", resolution);
-		SetUniformf("resolution_environmentMap", EnvironmentMapConfig::environmentMapResolution);
+		SetUniformf("resolution_environmentMap", (float)EnvironmentMapConfig::environmentMapResolution);
 	}
 };
